--- counter-wespeak/vercel.json
@@ -1,8 +1,0 @@
-{
-  "cron": [
-    {
-      "path": "/api/resetCounter",
-      "schedule": "*/2 * * * *"
-    }
-  ]
-}
